Consolidate HudsonPrivateSecurityRealmTest & HudsonPrivateSecurityRealm2Test.
